+++ packages/components/alert/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/alert
 
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/typography@0.0.33
+
 ## 0.0.36
 
 ### Patch Changes